[Perf pipe] Initial plotting function
This is an initial function to display a plot for the Sched PIPE test.
The plot report also an "unfairness index" (i.e. [0,1]), the higher the value
the more unfair the scheduler has been measured to be in the target system

.:: Unfariness index
The bigger the difference between the run-time and the average cycle-time,
the more unfair the scheduler has been.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>