[DumpTable] Add parsing of testing configuration
All the parameters of a test are saved into a corresponding ".txt" test
description file. This patch parse such a file in order to recover some
test information which could be used to properly dump tablular data.

For example, the test description file reports the FTrace events collected
during the test. Such events could be used to enable or not the production
of certain tables.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>