[DumpTables] Update latency filter to report the CPU number
Each time a latency/slice is reported we dump also the CPU on which the
task has been executed. The reported CPU could be "shifted" by a cpu_base
constant which could be used to report the actual CPU on systems using
the IKS cluster switcher on big.LITTLE ARM platforms.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>