[DumpTables] Add IKS support on latency filter
Once running on a big.LITTLE ARM system, with IKS support, the kswitcher
kernel thread is in charge to switch among big and LITTLE clusters.
This exploits events generated by the kernel switcher to properly set the
cpu_base in such a way to report the actual CPU on witch a certain switch has
happened.

NOTE: this patch is fine only for IKS-Lite implementations where all the CPUs
are switched at the same time. In this case we still have the possibility to
have some switch events reported in the new CPUs even when they are just going
to switch, since we are switching the cpu_base once CPU0 has switched.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>