[Test] Ensure to empty the trace buffer on every start
Each time a new trace is stared we empty the buffer and set it to 10MB per
CPU, which should be enough to capture all the interesting events for
reasonable benchmark use-cases.
NOTE: on Chromebook v2 we have more than 630M LowMemory available when
running a standard IKS kernel, thus 80MB for trace events should not be an
issue.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>