[Test] Add support to trace IKS scheduler
The IKS testing requires to force the usage of the LITTLE cluster just before
to start the trace since the trace parsing scripts assumes that the system
is initially running on that cluster.

This patch force a "powersave" governor on CPUFreq just before to start the
trace, which ensures to be on the LITTLE cluster at the beginning of a trace.
Right after the trace has been started we switch back to the chromebook
default interactive governor.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>